rename db client
